chore: Update debug flag handling in server.py
This commit updates the handling of the debug flag in the server.py file. Instead of checking if the debug flag is enabled, it now checks if the debug flag is set to "yes" or if it is a boolean value of True. This change ensures that the debug mode is activated correctly based on the configuration settings.